refactor(COR-1811): Refactored variants components
