+++ packages/app/src/domain/variants/data-selection/get-variant-bar-chart-data.ts
@@ -1,0 +1,3 @@
+export function getVariantBarChartData() {
+  return null;
+}
