+++ packages/app/src/domain/variants/data-selection/index.ts
@@ -1,3 +1,4 @@
+export * from './get-variant-bar-chart-data';
 export * from './get-archived-variant-chart-data';
 export * from './get-variant-order-colors';
 export * from './get-variant-table-data';
