--- packages/app/src/domain/variants/variants-stacked-area-tile/index.ts
@@ -1,1 +1,0 @@
-export { VariantsStackedAreaTile } from './variants-stacked-area-tile';
