+++ packages/app/src/domain/variants/variants-stacked-bar-chart-tile.tsx
@@ -1,0 +1,11 @@
+import { MetadataProps } from '~/components';
+
+interface VariantsStackedBarChartTileProps {
+  title: string;
+  description: string;
+  metadata: MetadataProps;
+}
+
+export const VariantsStackedBarChartTile = (_props: VariantsStackedBarChartTileProps) => {
+  return <></>;
+};
