+++ packages/app/src/domain/variants/variants-table-tile/components/variants-table.tsx
@@ -1,8 +1,8 @@
-import { VariantRow } from '~/domain/variants/static-props';
 import { useBreakpoints } from '~/utils/use-breakpoints';
 import { TableText } from '../types';
 import { NarrowVariantsTable } from './narrow-variants-table';
 import { WideVariantsTable } from './wide-variants-table';
+import { VariantRow } from '~/domain/variants/data-selection/types';
 
 type VariantsTableProps = {
   rows: VariantRow[];
@@ -12,13 +12,5 @@ type VariantsTableProps = {
 export function VariantsTable({ rows, text }: VariantsTableProps) {
   const breakpoints = useBreakpoints();
 
-  return (
-    <>
-      {breakpoints.sm ? (
-        <WideVariantsTable rows={rows} text={text} />
-      ) : (
-        <NarrowVariantsTable rows={rows} text={text} />
-      )}
-    </>
-  );
+  return <>{breakpoints.sm ? <WideVariantsTable rows={rows} text={text} /> : <NarrowVariantsTable rows={rows} text={text} />}</>;
 }
