--- packages/app/src/domain/variants/variants-table-tile/index.ts
@@ -1,1 +1,0 @@
-export * from './variants-table-tile';
